add basic message data structure
I've copied the basic message data structure from fullnode, which is
designed for bitcoin. It is undecided whether we should use a message
structure similar to bitcoin or to use something different. However,
this gives us a place to start.

I've also added some tests that show the thing works as promised both in
node (or rather, io.js) and web browsers.

First of all, make sure you are running io.js. The easiest way to do
that is to install the latest version of io.js directly. The better way
is to use nvm (node version manager), and then install io.js with "nvm
install io.js", and then type in "nvm use" in the datt directory. That
will use the version of node recommended by datt, which is io.js.

In order to run the tests in node, type "npm run test-node". In order to
run them in the browser via karma, type "npm run test-browser". You will
need both Firefox and Chrome installed for that to work, since the tests
are ran in both. In order to run the tests in the browser with mocha
directly, which is useful for pinning down and re-running specific tests
over and over, first run "gulp build-tests" and then run the test
facilitator app with "node bin/testapp.js" and then open up
"localhost:3000" in your browser.

Note that for browsers the intent is that the datt app is built as
datt.js. Notice how it requires only part of the fullnode library, but
not the whole thing, by digging into the fullnode library and accessing
the files directly. This is better to improve the filesize of the
resulting browser build. We do not necessarily need to use all of
fullnode - only the pieces that are relevant to datt (although that will
probably be most of fullnode).