fix "js standard" issues
There were a variety of complaints from "js standard" about the code
formatting. I fixed all the issues in lib/*.js. I did not fix all the
issues in the test/* directory, because those issues are all about
"describe" and "it" not being defined, and the jslint ability to ignore
that didn't seem to work.