switch from bitcoinjs-lib to bitcore
bitcore has a number of advantages over bitcoinjs-lib:

- It is faster, because it uses Fedor Indutny's ultra vast crypto libraries.

- It has more features, including payment channels (in a separate module).

- It's what webcoin uses, so if we use that, we save a lot of filesize.

I also detected one minor bug, which is that the private key big number was
being derived wrongly. I've updated the tests to reflect this. The bigi library
has a .fromHex() method which is what should have been used. It was assuming
the hex string being passed in was something other than hex, presumably base
10, and just giving a completely wrong result.