make storage interface asynchronous, add tests
IndexedDB and leveldb are both asynchronous, so our contentStore interface
should reflect that. I've made the putContent and getContent methods
asynchronous. They now return a promise. Since we're still using localStorage
in the browser, which is technically synchronous, this is not actually needed
yet, but will be when switching the storage to IndexedDB.

I've also added tests for putContent and getContent.