add AsyncCrypto interface - work towards threads
Cryptography is slow, and in order to make sure networking and UI works fluidly
without blocking, we need to use threads or separate processes to offload
computation. The AsyncCrypto interface has an asynchronous interface, based on
promises, for slow cryptographic operations. It is not complete yet and does
not yet support workers, but is an example of one way to abstract away the
cryptography into an asynchronous interface that, in principle, could be placed
in separate threads (web workers in a browser, child process forks in node).

This interface may not be the best strategy for putting cryptography in
threads. There is a module that may make this much easier for us:
https://github.com/josdejong/workerpool