update AsyncCrypto iface w/ default static methods
It used to be that you had to create a new AsyncCrypto instance to use
AsyncCrypto. Although it would reuse the same global workerpool, that was still
annoying. This updates things so you can just use static methods, which work as
expected, i.e. AsyncCrypto.sha256(data) returns a promise to the hash of data.