allow overriding global workers
You should be able to use the AsyncCrypto static methods for convenience, which
will automatically create a new workerpool if one does not already exist. If
for some reason you want to create your own worker pool for use with
AsyncCrypto, you can do that by passing in your workers in the constructors,
and your new AsyncCrypto instance will use that pool instead.