self -> this
Before the invention of "bind", it was necessary to define "var self = this" at
the top of functions in javascript so that when referencing "this" inside a
closure, you could use the variable "self", rather than "this", since "this"
would refer (counterintuitively) to the global object. However, with the
invention of bind in ES5, this is no longer necessary, since you can just bind
the function to the correct "this".