disable karma tests by default
We have issues with karma not working correctly. On my machine, karma
"disconnects" frequently, but not always, when the "contentStore" tests are
run, presumably because creating the database times out or because it slows
things down. I'm not really sure. I've tried increasing the timeout and
increasing the tolerance for disconnect, which helps a little. The disconnect
tolerance is especially helpful when set to a number like 5, because it will
try again under a disconnect and usually eventually works. But it's so slow
that it's not useful. We're better off running the mocha browser tests until we
can make the karma tests run without failing.

To run the mocha browser tests, try this:

gulp ./bin/testapp.js

Then open up localhost:3000/tests.html in your browser.