remove redundant workers
The purpose of AsyncCrypto was to run cryptography in a worker pool. The
purpose of Workers was to manage the worker pool. But because we're using the
workerpool library, Workers is basically pointless and doesn't do anything that
AsyncCrypto can't do. I have moved all of the Workers code into AsyncCrypto,
and removed Workers. The only reason we may ever want to revert this if there
are things other than crypto we want to put in workers. but then maybe we
should just rename AsyncCrypto rather than reintroducing Workers.