make Message not depend on Datt
We should have a hierarchical dependency tree, with no circular dependencies.
Circular dependencies make it very difficult to understand and debug code. I
have removed the implicit dependency on Datt from Message and made appropriate
changes to ContentDiscovery. We still need to update ContentDiscovery not to
depend on Datt. The solution there may be to move the ContentDiscovery code
into Datt.