Datt -> Main, datt -> main
the lib/datt.js file is the main entry point to the datt node application.
However, it is not appropriate to call this thing 'datt' - it's like calling a
car's engine 'car'. Just because the engine is the most important piece does
not mean it is the car. The car is composed of all the components of a car,
including the engine. Datt is composed of networking code, database code,
protocol code, user code, and so on, all of which constitute 'datt'.

The Main module now needs to be broken up into at least three distinct pieces:
- The connection manager
- The message handler
- The "main loop" which sets everything up and begins listening for things

And furthermore, the ContentDiscovery module needs to be refactored into this
new architecture.