disable app, edit html only, "own your content"
We were using both npm and bower to build the landing page. However, the
landing page is actually a very simple html page with some style and graphics.
I was having problems using bower (it kept timing out when trying to download
github packages, although github itself was working fine), so I decided to
simply use the rendered html page and not use bower or npm at all.

I have also changed the main title of the page from "Decentralie all the
things" to "Own your content", since I think the latter is the message we
really want to tell people. It's actually not necessary to define what "datt"
stands for most of the time.