re-enable app; works in node 4.0
I've managed to make the app work on my machine by fixing some firewall issue
that was breaking bower. I've also made the app work in node 4.0 by fixing some
minor issues, as well as removing the browser-sync dependency which seemed to
be broken in node 4.0, and is also not necessary.