restore AsyncCrypto with fullnode and babel
AsyncCrypto is an interface for doing cryptography in a web worker (browser) or
child process forks (node). There is a tool we are using called workerpool that
manages workers both in browsers and node. AsyncCrypto was introduced in an
earlier commit, but removed for the rewrite. Rather than use fullnode's
workers, I am restoring AsyncCrypto which is better suited to us. I have also
switched to using babel instead of es6ify because it seems to have become the
standard es6 -> es5 transpiler. I'm also using fullnode instead of bitcore,
because we are now using node 4.0 instead of 0.12, which supports ES2015, which
fullnode is written in.

The way to use AsyncCrypto is basically to noticed anytime you are doing
something with cryptography, particularly point multiplication (e.g., signing
and verifying) or hash functions, you should use AsyncCrypto for it. If the
particular crypto you want to use has not been implemented in AsyncCrypto, then
you should implement it. Never do crypto in the main thread, since it is slow
and blocking and will create a poor user experience.