convert AsyncCrypto to inject pattern
The inject pattern was implemented in fullnode and modeled after AngularJS's
pattern for injecting dependencies. It lets you easily create new "classes"
with different dependencies. It is a little cumbersome to wrap your classes
with it, but nice when you need to override the database, network, or
configuration in a module (or anything else).