add datt-node and correctly initialize it
The react app should create a new dattnode object, which has the logic of the
datt p2p protocol and database, and should initialize it. After it is
initialized, we update the state to note that it is initialized and display
that on the screeen. This now works as intended.