generate new User with AsyncCrypto
I've changed the .fromRandom() method of the user to use AsyncCrypto to
generate a new "identity", i.e. a BIP39 mnemonic along with BIP32 master xprv
and master xpub. When you randomly generate a user, it leaves the username
blank.  You can then set the user name with .setName(username). You can also
access .fromJSON and .toJSON to retrieve/store the user to a json object which
could be placed in a database.

Unlike fullnode, I've decided to make User aware that it is dealing with
cryptography, and thus needs to use the AsyncCrypto interface. This will
probably work fine for now - although I'm not sure that is the best
architecture. Another way to do it would be to create the user itself in a
worker, then transmit the object back to the main thread after it is created.
That is the architecture used in fullnode. Both have plusses and minusses. Not
sure what's best.