incompletely working display of user mnemonic
The changes to datt-node to create a new user is broken, because a web worker,
but the build process in datt-react is not creating the worker file. I think
the solution to this is just to merge the repos and have one unified build
process that builds everything. That would simplify things.