new initial commit: merge datt-node and datt-react
This repo was originally designed as the "main repo for datt" that would
contain the p2p logic and database (which would run both in node and a browser)
and a graphical front-end. However, when we decided to use Ember for the
front-end, we also decided to adopt the Ember architecture style of keeping the
front-end in a separate repository called datt.ui. So we created a new
repository, datt-node, which contained the p2p protocol. This repo, simply
"datt", was to bring in both datt-node and datt.ui into a single place where
users could run a node and see the interface.

However, after dealing with the three separate repositories, I realized that it
was very painful to maintain the build process and tests across the different
repos. I also decided to create a front-end in React (datt-react) rather than
Ember, which seemed simpler and well-suited the style of application we are
building (where all the logic is in the "node" part and the front-end basically
just needs to display the state of the node). Along the way I also began a
complete re-write of datt-node which I realized would be much easier to write
if paired with the front-end.

So I've now merged datt-node and datt-react together into the datt repo with a
unified build process. The lib/* files are the logic of the application and the
view/* files are the graphical front-end in React.

Note that this new datt repo can still be used with the Ember front-end if
desired. Just require datt and then use the p2p logic, for which the entry
point is lib/index.js.

Things are working as well as they were before, but there is still a huge
amount of work to do. The current status is:

- The react front-end loads the node, but current fails to correctly generate a
  new user due to a broken build process (which will be easier to fix now that
  things are merged).

- The node does not contain any p2p logic and ony has a minimal database. But
  this should also be easier to build and iterate with in the new repo.