properly initialize and display new user
The front-end now correctly generates a new user using dattnode and displays
that user's mnemonic. Right now the user is not saved to the database - but the
way this should work is that a "new user" is only generated if one was not
already stored in the database.

There is a somewhat sophisticated build and display process going on that
deserves to be explained.

For one, we are simply delivering a static html file from build/index.html that
loads the javascript. There are two javascript files being loaded (for now):
datt-node.js and datt-react.js. datt-node loads the global DattNode object,
which you can use to create a dattnode instance, which will then initialize the
database and p2p connections (though only the database works at present).

The datt-react.js file is the react app built to a single file. It is separated
from datt-node.js both to make it straightforward to use this build with other
front-ends, and also to make managing workers simpler.

Two other files, datt-node-worker.js and datt-node-workerpool.js are loaded by
datt-node.js, and so they are included in the build. Normally there would be
just one worker file, but because we are using the workerpool module, it
requires the use of a separate workerpool file. So all-in-all there are 4
javascript files being loaded:

- datt-node.js - database and p2p connections
- datt-react.js - react front-end
- datt-worker.js - cryptography, executed in worker
- datt-workerpool.js - workerpool management code

Basically, these files are loaded from:

- lib/*.js => datt-node.js, datt-worker.js, datt-workerpool.js
- views/*.jsx => datt-react.js

Note that the "datt node" can run in node.js, and uses a separate database and
communication layer (albeit it is incomplete at the moment).