add watch-test-node script
It is useful to be able to run the tests continuously while developing. You can
now do that with "npm run watch-test-node" which runs the node tests over and
over again upon file changes.

There are some other uses of gulp-watch that we should add next, particularly
rebuilding the browser bundles and refreshing the page.