add karma "run once" tests
In order to speed up development, karma can help automatically run the browser
tests. We had a lot of issues with karma in datt-node before for some reason,
which I suspect has to do with a combination of PouchDB and Chrome. I noticed
the same problems this time. So rather than run Chrome, I run Firefox, and
things seem to run smoothly.