rename datt-node to datt-core
There are too many things named "node". We have node.js, fullnode, a bitcoin
full node, and datt-node. Although datt-node is appropriately named as a node
on a p2p network, it is also appropriately called core, i.e. the core of the
datt application. All the variables and files now refer to this as datt-core,
which so it should be less confusing.