make test build depend on bundle
The bundle, i.e. datt-core.js, contains the polyfill, which must be included in
browser-side tests, so it is proper to build the bundle when building the
tests. Hypothetically someone may build and eexecute the tests without having
had previously built the bundle, and then the tests would fail - this fixes
that.