browser-sync
You can now run "gulp test-serve" or "npm run test-serve" to run a browser-sync
server which will watch for updated js files and will reload your browser. The
way it works is by inserting a new js file into the html that opens a socket
and reloads files.

This is probably better than bin/testapp.js for most purposes. The way to use
it is just to run "gulp test-serve" then open up localhost:3000/index.html
and/or localhost:3000/tests.html and start editing code. They will refresh when
things finish building.

There is only one problem, which is that the build process is slow, so your
browser won't refresh very quickly. This can be solved by using browserify
watchify, which I have not yet bothered to include, but that is high on my list
of TODOs.