add log about browserSync to testapp
bin/testapp.js is what we used to use to run a test server that would deliver
the static files to the browser. However, this is now duplicated by
browserSync, which you can run with either "npm run serve" or "gulp serve".
browserSync is strictly better than our test app for development purposes,
because it does all the same things, plus automatically syncs changes. However,
I don't want to delete testapp.js, because there will come a time when it is
relevant again. browserSync is designed only to serve static files - it is not
designed to run a real server with back-end logic. That is what testapp.js will
morph into. Therefore, testapp.js will stay around for now, and just now be
used.