use open sans; watch css files for changes
Chris' front-end designs use Open Sans, so I've included that as our main font.
Note there is one potential issue, which is that I'm linking to Google's
servers to get the font file. This may be undesirable from a security and
privacy point-of-view. However, it will almost certainly load faster for
people.