add style; fix gulp watch
The main index page now has a bit of style thanks to twitter bootstrap. The
page now has structure, where there is a title bar, a large panel on the left
for the content, and another panel on the right for user information.

I had to spend a lot of time fixing the build. I have a feeling build issues
are going to be a continuous nightmare. So it goes with any complex project.
The issue was that although things appeared to be rebuilding upon watch, they
weren't actually. Turns out I was misusing watchify. The way it should work is
that the only method you call after files change is .bundle() and then write
the bundle. There were a bunch of other issues around catching errors that I
also had to fix. Plus one especially irritating issue where after noticing a
file has changed, if you try to create a bundle immediately, you will get a
"file doesn't exist" error. So you have to wait hundreds of milliseconds before
trying to creat the bundle. After playing around a lot, it seems 1000 ms seems
to solve the issue, which seems very high to me. But 750 ms wasn't enough.

All-in-all, the build nightmare probably isn't over yet, but I will try to
proceed and perhaps get a a point in the future where I understand how to build
well enough that I can rewrite the build proess.