simplify gulpfile slightly
There is a problem with gulp watch where for some reason I don't yet
understand, sometimes gulp watch breaks and stops showing changes. This has
something to do with browserify and errors that aren't being handled correctly
- not really sure. In any case, in the course of investigating the issue I've
simplified the gulpfile slightly. There is no reason to have the "called"
variables if we just take the initial browserify code out of the build
functions.