allow setting username (in memory only)
You can now use a form in the user box to set your username. No signing or
authentication is done, although it must pass the validation in the setUserName
function or it will silently throw an error and not update. Some next steps
here are to make it look a little nicer by using twitter bootstrap's form
groups, actually signing the name, and storing it to the database somehow.