reset browser-sync to 2.9.6
I had weird issues with browser-sync 2.9.7 where the server would try to load
2.9.6, which didn't exist. This is fixed by switching back to 2.9.6. This may
be a peculiarity with my local version, or it may be a bug in 2.9.7. Either
way, things seem to be fixed with 2.9.6.

On a related note, I updated npm and was surprised to to discover that I am now
on npm 3 instead of npm 2. After a few weird anomalies involving not-found
dependencies, after a few rounds of "npm install" things seem to work just fine
in npm 3.