add msg from fullnode
I've copied the bitcoin p2p message structure by copying it directly from
fullnode. This works for a first step, but the p2p message structure of bitcoin
is not designed for datt. We should update the p2p structure to be the same as
what's suggested in p2p.md in dattdocs.

The only change I've made so far is to limit the message size to 100 kb instead
of 32 mb. 100 kb should be sufficient for text messages. If we ever need to
transmit more, we can put data in parts.

Note that I've left the "magicbytes" the exact same as bitcoin. We may want to
change that along with the message structure.