get rid of Msg checksum
The format of Msg was copied from bitcoin. However, we are almost certainly
going to be using authenticated "low-level" network protocols such as secure
web sockets and web RTC. We therefore do not need a checksum on our messages,
because those protocols already include a checksum. Thus we can speed things up
slightly by getting rid of the checksum.