update to browser-sync 2.9.7 (again)
I figured out the issue with browser-sync. It is caching files somewhere so it
doesn't have to inject them again. It was storing a cached version of
index.html somewhere, and to get it to regenerate this version, I just did
"touch index.html", which signaled to browser-sync not to use the cached
version. What this means is that browser-sync now injects the correct
browser-sync javascript version into the source code. So if you have the same
problem with some HTML files where your browser console reports that the wrong
version of browser-sync is being loaded and is not found, try running "touch
myfile.html" and see if that fixes it.