use asynchronous gets
There is a good reason to always use asynchronous methods on dattcore, even if
they're not necessary. We're assuming for now that you are controlling a node
that is running in your web browser. But we will need to add support for
controlling remote nodes. That is, you might want to run an interface locally,
but log in to a node you are running somewhere else to view its state and issue
commands. This will be most useful for remotely administering nodes that are
permanently hosted somewhere. We don't need to add support for that yet, but so
long as we make all methods asynhronous, we will be able to add support for
remotely-executed calls to dattcore at some later time. For instance, suppose
you are running a node, and you want to see what data it is storing, and
perhaps delete some of the data, via the same graphical interface as normal. We
will be able to add support for that without rearchitecting everything if we
merely make sure that all calls are asynchronous, even if they don't need to be
when running locally.