add timeout for watch-test-node
I occassionally get "file not found" errors when re-running the tests which, as
explained in earlier commits, has something to do with watching files that have
just changed, and it seems to be fixed by just adding a small timeout. In this
case, 500ms seems to work, at least as best I can tell right now.