slightly improved error handling in gulpfile
There was a problem that when you produce a syntax error in a .jsx file, the
watchify build in "gulp serve" breaks and you can't keep refreshing and must
fix the error, and then restart "gulp serve" to be able to see the changes.
I've fixed this by erasing and re-creating the bundle upon an error. This seems
to fix things ... sometimes. I can produce a syntax error, fix it, and then
things refresh and work now.

However, this approach is not really desirable because I have to recreate the
watchify object. This works but I believe it is slow. Also, it still seems to
work inconsistently. I will have to continue investigating.