add verifyCompactSig
There are two types of signatures used in bitcoin, which we can adopt for other
purposes in Datt. I have now added verifying "compact" signatures, which are
signatures that also contain a recovery factor that lets you easily derive a
public key. We should also add signing which produces a compact signature.

Compact formatted signatures have one very nice property: They are always
exactly 65 bytes. That is unlike DER signatures, which vary in size. Always
being exactly the same number of bytes is nice for transmitting on a network or
storing, since you can be sure how many bytes it must be.