auth message
The auth message is what you make when you want to prove your identity. There
is a specific reason why a block hash is included in this message. It will
often be the case that you want to authentiate with someone and prove that you
actually have a private key right now, i.e. because you are signing a new
message. Signing the latest block hash is a way of proving you presently posses
the private key, and are not just relaying some stale signature from someone
else.