make views modular; add a test
I'm working on making the front-end thoroughly tested. One step is to put the
components into their own files so they can be individually tested. Another is
to make sure the basics of testing them actually works. We already had the test
framework working with the front-end, but there were no actual tests other than
loading the layout file. I've now added a basic test to the layout to make sure
that it is actually rendering. I now need to go add tests for each of the
modules.