remove injector
The injector was a bit annoying to use and doesn't seem to be useful at this
time. If we ever get to a situation where we need to regularly mock
dependencies, we can either add it back, or we can find another injector
suitable for datt. But for now, it was causing me too much pain to think about.
Basically, all the problems that it might be able to solve are better solved
with normal configuration at this time.

It's worth noting that this repo originally started with the injector, because
I copied the style of fullnode. fullnode makes extensive use of the injector
for distinguishing between mainnet and testnet. Then we opted not to use it for
datt-node, because it was too cumbersome for other people to learn it. Then
when I started rewriting datt-node, I put it back in again, in the name of
consistency betwen this project and fullnode. Now I'm taking it out again
because it's too cumbersome to deal with. Such is life.