pass dattcore_status as a prop
The rendering of a box may depend on whether dattcore has been initialized or
not. We can render the user box with a blank mnemonic before the mnemonic is
available, but we would like to update the state after dattcore has been
initialized, because then the real mnemonic will be available. There are
multiple ways to handle this - perhaps the user box could notice an event from
dattcore, and update its own state when the "mnemonic has been updated" event
is received. For now we just pass the state of dattcore to the boxes.