make another react test work
You can't just make a react object and start calling its method - I don't know
why. But React has a built-in test framework that you can use. Unfortunatey
this framework depends on the DOM. So you can make this thing work in node with
a virtual DOM. I've added one test that seems to work both in node and a
browser. It works out of the box in a browser. For node, I had to use a virtual
DOM. I also had to exclude this virtual DOM from the browserify tests since it
is written in es6 and messes with firefox, even though it doesn't actually run.