content auth
ContentAuth means authenticated content, i.e. content that has been signed.
Actually, not just signed, but we have also added the block hash, block height,
untrusted date, and a bitcoin address to the data before signing it. This is
basically saying "the owner of this public key asserts that at this block
height (x) I have signed this data, and here is a bitcoin address you can use
to send me money if you want." This proves that the signature is new (because
you can't sign an unknown future block hash).