add async methods which are safe for main thread
Frankly, I'm not sure at an architectural level how best to handle all the
crypto that must be asynchronous and run in worker. I suppose it's fine if the
AsyncCrypto interface is universal and we can use it from within any class.
Seems reasonable to have synchronous, blocking versions of crypto methods
alongside asynchronous, non-blocking methods. This allows us to create the
simple synchronous methods and test then and then create slightly more
complicated asynchronous versions and test those and ensure the results are the
same. It also means we can write all "logical" methods where they belong - in
the class that logically owns them.

I'm also not really sure how best to distinguish the names of synchronous
methods from asynchronous equivalents. In fullnode, I used the prefix "await"
for some of these, which I later replaced with "wait". In this case, seems
"async" seems to capture the essence of what I'm trying to accomplish without
being too verbose. So "async" it is, for now.