DBContentAuth: save and get authed content from db
I'm assuming there is no reason to save unauthenteicated content, and the only
type of content you will actually want to store is authenticated content, i.e.
ContentAuth. As such I have created db-content-auth or DBContentAuth which is
an interface to the database for saving and getting ContentAuth objects.

Like with workers, the interface assumes we're using buffers, but the things
that are actually inserted are hex strings. That is because hex strings are
what works with pouchdb (just like hex strings are what works with our workers
library), but we will want to update this in the future to simply be buffers,
as converting to/from hex strings is pointless computation that slows things
down.