CoreBitcoin
This will eventually be the "bitcoin management" class. There will not be
bitcoin logic in here, but this object on dattcore will manage other objects
that do have logic and peer connections. For instance, there may be a wallet
object, and/or an unspents object, and/or a blockchain object.

Closes #45