get/set -> asyncGet/asyncSet
Traditionally, get and set methods are synchronous, that is they immediately
set the value and either return a value or return this. Asynchronous methods
don't, and the way you use these methods must be different (basically, if a
method is asyncronous, your containing code must also be asynchronous). It was
inconsistent and annoying to use "set" and "get" for both synchronous and
asynchronous methods. It seems better to prefix asynchronous methods with
"async" - that way you know a get or set method is async without reading the
source code or comments, and can use it appropriately.