content submit form actually posts to db
I'm still not sure how to properly test stuff with React, so I'm submitting
this without tests, however I was able to confirm by hand that the content was
actually being submitted to the database. The front-end right now should be
regarded as a sort of testing canvas which will need to be mostly rewritten,
since I haven't put much thought into the architecture or the testing yet. It's
just a way to graphically ensure that dattcore is working. We will probably
want to redo all or most of the front-end once the prototype is finished.