LICENSE file
We're using the MIT license as stated in the package.json since the beginning.
The only tricky part about adding the MIT LICENSE file in here at this stage is
that we have to specify who to give the copyright to. Seems the fairest thing
to do for now is to give copyright to "Datt Developers". At some later date we
will probably add Datt, Inc. to this file.