decrease max size from 1mb to 30kb
I thought I had set the max size to 100kb, but evidently I had it set to 1mb.
This shouldn't be very high as it is a way for nodes to attack each other. It
simply needs to be large enough for most messages, and no larger. I checked the
datt master plan, which is presently about 25kb, and that is almost certainly
longer than almost all content that will be posted to datt. So by setting this
to 30kb, we should be able to post the datt master plan to datt with a little
bit of room to spare. That seems like a reasonable benchmark.