update to react 0.14
There is a new version of React. I was able to update easily except the
ReactDOMServer.renderToString method was giving me a weird warning about how I
was using more than one instance of React - and the warning was there both in
node and a browser. I don't know whether this is my mistake somewhere or
whether this is a bug in React (I'm certainly not intentionally including React
twice). In any case, I fixed the issue by changing the way that test works.
Right now I'm not using ReactDOMServer at all and I'm not using renderToString
at all. That seems to fix the issue.

See Facebook's launch announcement about React 0.14:
- http://facebook.github.io/react/blog/2015/10/07/react-v0.14.html