get/save -> asyncGet/asyncSave
Our convention is that all methods that return a promise are prefixed with
"async". Although this is verbose, I believe the consistency is worth it. It is
very nice knowing whether a methods returns a promise or not by its prefix.
That only works if every method is prefixed correctly - even database get/save
methods which are (obviously) async.