init -> asyncInitialize
As mentioned in the previous commit, although it increases verbosity, I believe
it is better for consistency that all methods that return a promise are
prefixed with "async", and that goes even for the "init" method of the various
objects. Our convention is that "initialized" is used as a synchronous
initialization method to set the default values of variables -
"asyncInitialize" is done when you want to "boot up" something such as network
connetions or the database which requires as async operation and therefore
returns a promise. You do not normally call initialize() by hand, because that
is done in the constructor. But you do call asyncInitialize() by hand when you
desire to boot up something.