NetworkBrowserWebRTC and PeerJS
There is now a basic "network" class for creating a "Web RTC network". In the
future there will be other classes for web sockets, some of which will be
specifically for node.js and some for browsers. This is a very weird way to do
things, but unfortunately it seems like the best way to handle the multiple
different types of networks we have to support. I tried using feross'
simplepeer, which theoretically works both in node and a browser, but
unfortunately it depends on wrtc, a node.js package for Web RTC, and that has
not yet been updated to work in node 4.0. So I've reverted back to the plan of
using PeerJS for our network connections in a browser. Browsers will then
connect to servers using web sockets, and servers will connect to each other
using web sockets. This creates a bunch of weirdness for handling the network
connections that I don't know how to avoid. I think no matter what we do,
having a p2p network that spans web browsers and servers is going to be weird
and complicated. I'm doing the best I can to manage that weirdness.

The philosophy of this change is as follows:

- There will ultimately be a CorePeers class that manages all the network
  complexity.

- CorePeers will depend on DBPeers (for storing peer info in the database) and
  Peers, which is where most of the logic will be.

- Peers, in turn, will use various Network classes. In node, we will use
  NetworkNodeSocket to create and receive connections. In a browser, we will
  use NetworkBrowserSocket to create connections (to servers) and
  NetworkBrowserWebRTC to make and receive connections to other browsers.

- The network classes in turn will have various connection classes, such as
  ConnectionBrowserWebRTC. You use a Network to create and receive connections.
  When a connection is established, you will have an instance of one of the
  Connection classes.

- This is reminiscent of how things work in fullnode, except fullnode uses TCP
  instead of sockets, and socket support is planned later to bridge them. With
  Datt, we don't need the TCP part, so things will actually be slightly
  simpler.

This commit adds just NetworkBrowserWebRTC, and various changest to our
gulpfile and testapp to make things actually run and testable. That is because
we now must run a PeerJS server to handle the rendezvous part.