ping and pong messages betwix connections
I have added automatic "pong" responses to connections, so that when a ping is
received, a connection will automatically respond with a pong. I should
probably move this higher up in the application stack - the Peers class is
probably what should handle all messages, including ping/pong, but it hasn't
been written yet. So for now I'll leave this here, then when Peers is ready, I
will move the ping/pong code there.