Peers - for managing p2p connections
As mentioned in multiple commits and comments, our network code unfortunately
must be kind of weird and complex because we need to support both web socket
connections in node and browsers and Web RTC connections in node and browsers.
There are a few catches with this:
- In node, web sockets can create and receive connections
- In node, web RTC will not work at all for the time being, although support
  can be added later in theory
- In browsers, web sockets can create but cannot receive connections
- In browsers, web RTC can create and receive connections

The Peers class is designed to manage this complexity. Under the hood different
"network" and "connection" types are being used. But someone who uses Peers
doesn't care - they just directly connect to someone given their
connectionInfo, or they receive a new connection.

Peers will not manage the database. The logic for what to do when a particular
message received therefore cannot live in Peers, because it would be incapable
of responding to a query to deliver content. Instead, that code will reside in
CorePeers, which has not yet been created.

Some TODO items:
- TODO: Add support for sockets in node and browsers
- TODO: Create CorePeers for managing the actual messages
- TODO: Set up listeners to Peers knows when a message is received and when a
  connection is made, etc.
- TODO: Hook up the UI so content can be sent from the UI and received content
  is displayed.